update:inbound query
