+++ WMS/src/main/java/wms/menu/controller/MainMenuController.java
@@ -1,0 +1,16 @@
+package wms.menu.controller;
+
+import wms.menu.menuView.LoginView;
+
+import wms.menu.menuView.MainMenuView;
+
+public class MainMenuController {
+
+    public void loginView() {
+        new LoginView().loginView();
+    }
+
+    public void mainMenuView() {
+        new MainMenuView().mainView();
+    }
+}
