--- WMS/src/main/java/wms/menu/controller/mainMenuController.java
@@ -1,5 +1,0 @@
-package wms.menu.controller;
-
-public class mainMenuController {
-
-}
