+++ WMS/src/main/java/wms/menu/menuView/ProductView.java
@@ -1,5 +1,6 @@
 package wms.menu.menuView;
 
+import wms.menu.controller.MainMenuController;
 import wms.menu.controller.ProductCategoryController;
 import wms.menu.controller.ProductController;
 import wms.menu.model.dto.ManufacturerDto;
@@ -35,14 +36,14 @@ public class ProductView {
     private Scanner sc = new Scanner(System.in);
     private ProductController productController = new ProductController();
     private ProductCategoryController productCategoryController = new ProductCategoryController();
-
+    private MainMenuController mainMenuController=new MainMenuController();
     // 상품 관리 화면
     public void productMainMenu() {
         String menu = """
                 =======================
                 1. 상품 관리
                 2. 상품 카테고리 관리
-                0. 나가기
+                0. 메인 메뉴
                 =======================
                 입력 : """;
         while(true) {
@@ -52,7 +53,7 @@ public class ProductView {
                 case "1" -> productController.productManage(showProductManage());
                 case "2" -> productCategoryController.productCategoryManage(showProductCategoryManage());
                 case "0" -> {
-                    return;
+                    mainMenuController.mainMenuView();
                 }
                 default ->{
 //                    if (!choice.equals("1") && !choice.equals("2") && !choice.equals("0"))
