--- WMS/src/main/java/wms/menu/menuView/mainMenuView.java
@@ -1,4 +1,0 @@
-package wms.menu.menuView;
-
-public class mainMenuView {
-}
