+++ WMS/src/main/resources/mybatis-config.xml
@@ -18,9 +18,14 @@
             <dataSource type="POOLED">
                 <property name="driver" value="com.mysql.cj.jdbc.Driver"/>
                 <!--아직 menudb같은 공간을 만들지 않아 3306/스키마 이름을 써야 합니다-->
+                <!--로컬 저장소 주소 아이디 비밀번호-->
                 <property name="url" value="jdbc:mysql://localhost:3306/hhourdb"/>
                 <property name="username" value="sh"/>
                 <property name="password" value="sh"/>
+                <!--techbrew 저장소 아이디 비밀번호-->
+                <property name="url" value="jdbc:mysql://techbrew4.iptime.org:3306/happyhourdb"/>
+                <property name="username" value="hhdb"/>
+                <property name="password" value="hhdb"/>
 
           </dataSource>
